refactor(mock-file): camelCase
