+++ boilerplate/test/setup.ts
@@ -1,7 +1,7 @@
 // we always make sure 'react-native' gets included first
 import * as ReactNative from "react-native"
 import mockAsyncStorage from "@react-native-async-storage/async-storage/jest/async-storage-mock"
-import mockFile from "./mock-file"
+import mockFile from "./mockFile"
 
 // libraries to mock
 jest.doMock("react-native", () => {
